Move CloneCraftFrame to it's own class file
